Create basic package structure
