+++ .gitignore
@@ -1,0 +1,8 @@
+# python
+venv
+
+__pycache__
+
+# Visual Studio Code
+
+.vscode
