+++ dm_assist/__main__.py
@@ -1,0 +1,5 @@
+from . import start
+
+# __main__.py is run when the package is run directly: python dm_assist
+
+start()
