+++ start.py
@@ -1,0 +1,4 @@
+import dm_assist
+
+
+dm_assist.start()
